property callback mode beta.
