+++ examples/linux/light/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_executable(light_device light_device.c tc_iot_shadow_property.c ../tc_iot_client_config_opts.c)
+add_executable(light_device light_device.c tc_iot_shadow_property.c tc_iot_shadow_property_common.c ../tc_iot_client_config_opts.c)
 
 if(ENABLE_TLS)
     target_link_libraries(light_device tc_iot_suite tc_iot_common tc_iot_hal mbedtls jsmn MQTTPacketClient)
